requirements.txt
blinker==1.4
certifi==2023.5.7
charset-normalizer==3.1.0
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
docker==6.1.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
packaging==23.1
pandas==2.0.1
psycopg2-binary==2.9.6
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.30.0
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
wadllib==1.3.6
websocket-client==1.5.1
zipp==1.0.0
